+++ sherpa/astro/ui/tests/test_astro_ui.py
@@ -19,10 +19,17 @@
 
 import os
 import unittest
+import tempfile
+
+import numpy
+from numpy.testing import assert_allclose
+
 from sherpa.utils import SherpaTest, SherpaTestCase
 from sherpa.utils import test_data_missing, has_fits_support
 from sherpa.astro import ui
-import numpy
+from sherpa.data import Data1D
+from sherpa.astro.data import DataPHA
+
 import logging
 logger = logging.getLogger("sherpa")
 
@@ -409,6 +416,153 @@ class test_stats_ui(SherpaTestCase):
         self.assertEqual('chi2', stat1)
         self.assertEqual('chi2', stat12)
 
+
+@unittest.skipIf(not has_fits_support(),
+                 'need pycrates, astropy.io.fits, or pyfits')
+class test_save_arrays_base(SherpaTestCase):
+
+    # _colnames:  specify column names (True) or not
+    # _fits:      use FITS format for output?
+
+    _colnames = False
+    _fits = None
+
+    def _read_func(self, filename):
+        raise NotImplementedError
+
+    def save_arrays(self):
+        """Write out a small set of data using ui.save_arrays
+        and then read it back in, to check it was written out
+        correctly (or, at least, in a way that can be read back
+        in).
+        """
+
+        # It looks like the input arrays to `write_arrays` should be numpy
+        # arrays, so enforce that invariant.
+        a = numpy.asarray([1, 3, 9])
+        b = numpy.sqrt(numpy.asarray(a))
+        c = b * 0.1
+
+        if self._colnames:
+            fields = ["x", "yy", "z"]
+        else:
+            fields = None
+
+        ofh = tempfile.NamedTemporaryFile(suffix='sherpa_test')
+        ui.save_arrays(ofh.name, [a, b, c], fields=fields,
+                       ascii=not self._fits, clobber=True)
+
+        out = self._read_func(ofh.name)
+
+        rtol = 0
+        atol = 1e-5
+        self.assertIsInstance(out, Data1D)
+        self.assertEqual(out.name, ofh.name, msg="file name")
+        assert_allclose(out.x, a, rtol=rtol, atol=atol, err_msg="x column")
+        assert_allclose(out.y, b, rtol=rtol, atol=atol, err_msg="y column")
+        assert_allclose(out.staterror, c, rtol=rtol, atol=atol,
+                        err_msg="staterror")
+        self.assertIsNone(out.syserror, msg="syserror")
+
+
+@unittest.skipIf(not has_fits_support(),
+                 'need pycrates, astropy.io.fits, or pyfits')
+class test_save_arrays_nocols_FITS(test_save_arrays_base):
+
+    _fits = True
+
+    def _read_func(self, filename):
+        return ui.unpack_table(filename, ncols=3)
+
+    def test_save_arrays(self):
+        self.save_arrays()
+
+
+@unittest.skipIf(not has_fits_support(),
+                 'need pycrates, astropy.io.fits, or pyfits')
+class test_save_arrays_cols_FITS(test_save_arrays_nocols_FITS):
+
+    _colnames = True
+
+    def test_save_arrays(self):
+        self.save_arrays()
+
+
+@unittest.skipIf(not has_fits_support(),
+                 'need pycrates, astropy.io.fits, or pyfits')
+class test_save_arrays_nocols_ASCII(test_save_arrays_base):
+
+    _fits = False
+
+    def _read_func(self, filename):
+        return ui.unpack_ascii(filename, ncols=3)
+
+    def test_save_arrays(self):
+        self.save_arrays()
+
+
+@unittest.skipIf(not has_fits_support(),
+                 'need pycrates, astropy.io.fits, or pyfits')
+class test_save_arrays_cols_ASCII(test_save_arrays_nocols_ASCII):
+
+    _colnames = True
+
+    def test_save_arrays(self):
+        self.save_arrays()
+
+
+@unittest.skipIf(not has_fits_support(),
+                 'need pycrates, astropy.io.fits, or pyfits')
+@unittest.skipIf(test_data_missing(), 'required test data missing')
+class test_save_pha(SherpaTestCase):
+    """Write out a PHA data set as a FITS file."""
+
+    longMessage = True
+
+    def setUp(self):
+        # hide warning messages from file I/O
+        self._old_logger_level = logger.level
+        logger.setLevel(logging.ERROR)
+
+        self._id = 1
+        fname = self.make_path('ciao4.3/pha_intro/3c273.pi')
+        ui.load_pha(self._id, fname)
+        self._pha = ui.get_data(self._id)
+
+    def tearDown(self):
+        logger.setLevel(self._old_logger_level)
+
+    def testWrite(self):
+        ofh = tempfile.NamedTemporaryFile(suffix='sherpa_test')
+        ui.save_pha(self._id, ofh.name, ascii=False, clobber=True)
+
+        # limited checks
+        pha = ui.unpack_pha(ofh.name)
+        self.assertIsInstance(pha, DataPHA)
+
+        for key in ["channel", "counts"]:
+            newval = getattr(pha, key)
+            oldval = getattr(self._pha, key)
+            assert_allclose(oldval, newval, err_msg=key)
+
+        # at present grouping and quality are not written out
+
+        for key in ["exposure", "backscal", "areascal"]:
+            newval = getattr(pha, key)
+            oldval = getattr(self._pha, key)
+            self.assertAlmostEqual(oldval, newval, msg=key)
+
+        """
+        Since the original file has RMF and ARF, the units
+        are energy, but on write out these files are not
+        created/saved, so when read back in the PHA has no
+        ARF/RMF, and will have units=channel.
+
+        for key in ["units"]:
+            newval = getattr(pha, key)
+            oldval = getattr(self._pha, key)
+            self.assertEqual(oldval, newval, msg=key)
+        """
 
 if __name__ == '__main__':
